migrate handling over to database code
