+++ core/database.py
@@ -16,11 +16,15 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
 
+from __future__ import annotations
+
 import asyncio
 import datetime
 import logging
+import re
 from typing import TYPE_CHECKING, Any, Self
 
+import aiohttp
 import asyncpg
 
 from core import CONFIG
@@ -31,25 +35,113 @@ from .models import FileModel, PasteModel
 
 if TYPE_CHECKING:
     _Pool = asyncpg.Pool[asyncpg.Record]
+    from types_.config import Github
+    from types_.github import PostGist
 else:
     _Pool = asyncpg.Pool
 
-
-logger: logging.Logger = logging.getLogger(__name__)
+DISCORD_TOKEN_REGEX: re.Pattern[str] = re.compile(r"[a-zA-Z0-9_-]{23,28}\.[a-zA-Z0-9_-]{6,7}\.[a-zA-Z0-9_-]{27,}")
+LOGGER: logging.Logger = logging.getLogger(__name__)
 
 
 class Database:
     pool: _Pool
 
-    def __init__(self, *, dsn: str) -> None:
+    def __init__(self, *, dsn: str, session: aiohttp.ClientSession | None = None, github_config: Github | None) -> None:
         self._dsn: str = dsn
+        self.session: aiohttp.ClientSession | None = session
+        self._handling_tokens = bool(self.session and github_config)
+
+        if self._handling_tokens:
+            LOGGER.info("Will handle compromised discord info.")
+            assert github_config  # guarded by if here
+
+            self._gist_token = github_config["token"]
+            self._gist_timeout = github_config["timeout"]
+            # tokens bucket for gist posting: {paste_id: token\ntoken}
+            self.__tokens_bucket: dict[str, str] = {}
+            self.__token_lock = asyncio.Lock()
+            self.__token_task = asyncio.create_task(self._token_task())
 
     async def __aenter__(self) -> Self:
         await self.connect()
         return self
 
     async def __aexit__(self, *_: Any) -> None:
+        task: asyncio.Task[None] | None = getattr(self, "__token_task", None)
+        if task:
+            task.cancel()
+
         await self.close()
+
+    async def _token_task(self) -> None:
+        # won't run unless pre-reqs are met in __init__
+        while True:
+            if self.__tokens_bucket:
+                async with self.__token_lock:
+                    await self._post_gist_of_tokens()
+
+            await asyncio.sleep(self._gist_timeout)
+
+    def _handle_discord_tokens(self, *bodies: dict[str, str], paste_id: str) -> None:
+        formatted_bodies = "\n".join(b["content"] for b in bodies)
+
+        tokens = list(DISCORD_TOKEN_REGEX.finditer(formatted_bodies))
+
+        if not tokens:
+            return
+
+        LOGGER.info(
+            "Discord bot token located and added to token bucket. Current bucket size is: %s", len(self.__tokens_bucket)
+        )
+
+        tokens = "\n".join([m[0] for m in tokens])
+        self.__tokens_bucket[paste_id] = tokens
+
+    async def _post_gist_of_tokens(self) -> None:
+        assert self.session  # guarded in caller
+        json_payload: PostGist = {
+            "description": "MystBin found these Discord tokens in a public paste, and posted them here to invalidate them. If you intended to share these, please apply a password to the paste.",
+            "files": {},
+            "public": True,
+        }
+
+        github_headers = {
+            "Accept": "application/vnd.github+json",
+            "Authorization": f"Bearer {self._gist_token}",
+            "X-GitHub-Api-Version": "2022-11-28",
+        }
+
+        current_tokens = self.__tokens_bucket
+        self.__tokens_bucket = {}
+
+        for paste_id, tokens in current_tokens.items():
+            filename = str(datetime.datetime.now(datetime.UTC)) + f"__{paste_id}-tokens.txt"
+            json_payload["files"][filename] = {"content": tokens}
+
+        success = False
+
+        try:
+            async with self.session.post(
+                "https://api.github.com/gists", headers=github_headers, json=json_payload
+            ) as resp:
+                success = resp.ok
+
+                if not success:
+                    response_body = await resp.text()
+                    LOGGER.error(
+                        "Failed to create gist with token bucket with response status code %s and response body:\n\n%s",
+                        resp.status,
+                        response_body,
+                    )
+        except (aiohttp.ClientError, aiohttp.ClientOSError) as error:
+            success = False
+            LOGGER.error("Failed to handle gist creation due to a client or operating system error", exc_info=error)
+
+        if success:
+            LOGGER.info("Gist created and invalidated tokens from %s pastes.", len(current_tokens))
+        else:
+            self.__tokens_bucket.update(current_tokens)
 
     async def connect(self) -> None:
         try:
@@ -64,15 +156,15 @@ class Database:
             await pool.execute(fp.read())
 
         self.pool = pool
-        logger.info("Successfully connected to the database.")
+        LOGGER.info("Successfully connected to the database.")
 
     async def close(self) -> None:
         try:
             await asyncio.wait_for(self.pool.close(), timeout=10)
         except TimeoutError:
-            logger.warning("Failed to greacefully close the database connection...")
+            LOGGER.warning("Failed to greacefully close the database connection...")
         else:
-            logger.info("Successfully closed the database connection.")
+            LOGGER.info("Successfully closed the database connection.")
 
     async def fetch_paste(self, identifier: str, *, password: str | None) -> PasteModel | None:
         assert self.pool
@@ -167,7 +259,13 @@ class Database:
                     if row:
                         paste.files.append(FileModel(row))
 
-            return paste
+        if not password:
+            # if the user didn't provide a password (a public paste)
+            # we check for discord tokens
+            LOGGER.info("Located tokens")
+            self._handle_discord_tokens(*data["files"], paste_id=paste.id)
+
+        return paste
 
     async def fetch_paste_security(self, *, token: str) -> PasteModel | None:
         query: str = """SELECT * FROM pastes WHERE safety = $1"""
